edited init and install scripts

+++ install.sh
@@ -2,32 +2,37 @@
 set -e
 
 # username password and hostname
-echo "Has the user password, root password and hostname been changed"
-# TODO: yes/no prompt here
+echo "Has the user password, root password and hostname been changed (y/n)"
+read -p '>' answer
 
-# install all libraries needed
-sudo apt update
-sudo apt -y install python3-pip screen
-pip3 install setuptools
-pip3 install awscli==1.11.18
-pip3 install boto3
-pip3 install pyserial
+if [ "$answer" = "y" ]; then
+    # install all libraries needed
+    sudo apt update
+    sudo apt -y install python3-pip screen
+    pip3 install setuptools
+    pip3 install awscli==1.11.18
+    pip3 install boto3
+    pip3 install pyserial
 
-# copy bash script to root
-cp .bashrc ../
+    # copy bash script to root
+    cp .bashrc ../
 
-# copy the init script to /etc/init.d and add to default start-up scripts
-echo "copy init file"
-cp ~/atlas-pi/atlas-init.sh /etc/init.d/
-echo "chmod file"
-sudo chmod 755 /etc/init.d/atlas-init.sh
-#echo "chown flie"
-#sudo chown root:root /etc/init.d/atlas-init.sh
-echo "add to defaults"
-sudo update-rc.d atlas-init.sh defaults
+    # copy the init script to /etc/init.d and add to default start-up scripts
+    echo "copy init file"
+    cp ~/atlas-pi/atlas-init.sh /etc/init.d/
+    echo "chmod file"
+    sudo chmod 755 /etc/init.d/atlas-init.sh
+    #echo "chown file"
+    #sudo chown root:root /etc/init.d/atlas-init.sh
+    echo "add to defaults"
+    sudo update-rc.d atlas-init.sh defaults
 
-# configure the aws creds
-echo "run: aws configure"
+    # configure the aws creds
+    echo "run: aws configure"
 
-# reboot to set everything running
-echo "lastly reboot"
+    # reboot to set everything running
+    echo "lastly reboot"
+else
+    exit
+fi
+
